// ProyectoAngular/PrimerAngular/src/app/models/estudiantes.ts
export class Estudiantes {
    id: string = "";
    cedula: string = "";
    correoelectronico: string = "";
    telefono: string = "";
    telefonocelular: string = "";
    fechanacimiento: string = "";
    sexo: string = "";
    direccion: string = "";
    nombre: string = "";
    apellidopaterno: string = "";
    apellidomaterno: string = "";
    nacionalidad: string = "";
    idCarreras: string = "";
    usuario: string = "";
}